+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - [BUGFIX] Prevents duplicated note IDs (#842).
 - Added a required Proof-of-Work in the faucet to request tokens (#831).
 - Added an optional API key request parameter to skip PoW in faucet (#839).
+- Proof-of-Work difficulty is now adjustable based on the number of concurrent requests (#865).
 
 ## v0.8.2 (2025-05-04)